+++ core/src/main/java/router/fu/RegExMatchResult.java
@@ -1,0 +1,8 @@
+package router.fu;
+
+interface RegExMatchResult
+{
+  int getGroupCount();
+
+  String getGroup( int i );
+}