+++ core/src/main/java/router/fu/RegExMatcher.java
@@ -1,0 +1,10 @@
+package router.fu;
+
+import javax.annotation.Nonnull;
+import javax.annotation.Nullable;
+
+interface RegExMatcher
+{
+  @Nullable
+  RegExMatchResult match( @Nonnull String text );
+}