+++ processor/src/test/fixtures/expected/com/example/callback/RouterFu_CallbackWithParams6.java
@@ -1,6 +1,6 @@
 package com.example.callback;
 
-import elemental2.core.JsRegExp;
+import akasha.core.RegExp;
 import java.util.Arrays;
 import java.util.Collections;
 import java.util.HashMap;
@@ -22,7 +22,7 @@ import router.fu.Segment;
 public class RouterFu_CallbackWithParams6 extends CallbackWithParams6 implements CallbackWithParams6Service {
   private final Parameter $fu$_regionCode = new Parameter( "regionCode" );
 
-  private final Route $fu$_route_region = new Route( "region", new Segment[]{new Segment( "/regions/" ), new Segment( $fu$_regionCode ), }, new Parameter[]{$fu$_regionCode, }, new JsRegExp( "^/regions/((?:[a-zA-Z0-9\\-_\\$\\.\\+!\\*'\\(\\)\\,\\\"~]|%[A-F0-9]{2})+)$" ), ( location, route, parameters ) -> regionCallback( parameters, route, location ) );
+  private final Route $fu$_route_region = new Route( "region", new Segment[]{new Segment( "/regions/" ), new Segment( $fu$_regionCode ), }, new Parameter[]{$fu$_regionCode, }, new RegExp( "^/regions/((?:[a-zA-Z0-9\\-_\\$\\.\\+!\\*'\\(\\)\\,\\\"~]|%[A-F0-9]{2})+)$" ), ( location, route, parameters ) -> regionCallback( parameters, route, location ) );
 
   private final Router $fu$_router;