+++ processor/src/test/fixtures/expected/com/example/parameter/RouterFu_NonDefaultNameParameter.java
@@ -1,6 +1,6 @@
 package com.example.parameter;
 
-import elemental2.core.JsRegExp;
+import akasha.core.RegExp;
 import java.util.Arrays;
 import java.util.Collections;
 import java.util.HashMap;
@@ -21,9 +21,9 @@ import router.fu.Segment;
 
 @Generated("router.fu.processor.RouterProcessor")
 public class RouterFu_NonDefaultNameParameter extends NonDefaultNameParameter implements NonDefaultNameParameterService {
-  private final Parameter $fu$_regionCode_821487049 = new Parameter( "regionCode", new JsRegExp( "^[a-zA-Z]+$" ) );
+  private final Parameter $fu$_regionCode_821487049 = new Parameter( "regionCode", new RegExp( "^[a-zA-Z]+$" ) );
 
-  private final Route $fu$_route_region = new Route( "region", new Segment[]{new Segment( "/regions/" ), new Segment( $fu$_regionCode_821487049 ), }, new Parameter[]{$fu$_regionCode_821487049, }, new JsRegExp( "^/regions/((?:[a-zA-Z0-9\\-_\\$\\.\\+!\\*'\\(\\)\\,\\\"~]|%[A-F0-9]{2})+)$" ), ( location, route, parameters ) -> MatchResult.TERMINAL );
+  private final Route $fu$_route_region = new Route( "region", new Segment[]{new Segment( "/regions/" ), new Segment( $fu$_regionCode_821487049 ), }, new Parameter[]{$fu$_regionCode_821487049, }, new RegExp( "^/regions/((?:[a-zA-Z0-9\\-_\\$\\.\\+!\\*'\\(\\)\\,\\\"~]|%[A-F0-9]{2})+)$" ), ( location, route, parameters ) -> MatchResult.TERMINAL );
 
   private final Router $fu$_router;