+++ processor/src/test/fixtures/expected/com/example/parameter/RouterFu_URIEncodedParameter.java
@@ -1,6 +1,6 @@
 package com.example.parameter;
 
-import elemental2.core.JsRegExp;
+import akasha.core.RegExp;
 import java.util.Arrays;
 import java.util.Collections;
 import java.util.HashMap;
@@ -21,9 +21,9 @@ import router.fu.Segment;
 
 @Generated("router.fu.processor.RouterProcessor")
 public class RouterFu_URIEncodedParameter extends URIEncodedParameter implements URIEncodedParameterService {
-  private final Parameter $fu$_keyword_755992427 = new Parameter( "keyword", new JsRegExp( "^((?:[a-zA-Z0-9\\-_\\$\\.\\+!\\*'\\(\\)\\,\\\"~]|%[A-F0-9]{2})+)$" ) );
+  private final Parameter $fu$_keyword_755992427 = new Parameter( "keyword", new RegExp( "^((?:[a-zA-Z0-9\\-_\\$\\.\\+!\\*'\\(\\)\\,\\\"~]|%[A-F0-9]{2})+)$" ) );
 
-  private final Route $fu$_route_region = new Route( "region", new Segment[]{new Segment( "/regions/" ), new Segment( $fu$_keyword_755992427 ), }, new Parameter[]{$fu$_keyword_755992427, }, new JsRegExp( "^/regions/((?:[a-zA-Z0-9\\-_\\$\\.\\+!\\*'\\(\\)\\,\\\"~]|%[A-F0-9]{2})+)$" ), ( location, route, parameters ) -> MatchResult.TERMINAL );
+  private final Route $fu$_route_region = new Route( "region", new Segment[]{new Segment( "/regions/" ), new Segment( $fu$_keyword_755992427 ), }, new Parameter[]{$fu$_keyword_755992427, }, new RegExp( "^/regions/((?:[a-zA-Z0-9\\-_\\$\\.\\+!\\*'\\(\\)\\,\\\"~]|%[A-F0-9]{2})+)$" ), ( location, route, parameters ) -> MatchResult.TERMINAL );
 
   private final Router $fu$_router;