+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 * Upgrade the `au.com.stocksoftware.idea.codestyle` artifact to version `1.17`.
 * Upgrade the `org.realityforge.org.jetbrains.annotations` artifact to version `1.7.0`.
 * Upgrade the `org.realityforge.arez` artifacts to version `0.193`.
+* Migrate from using Elemental2 to Akasha when interacting with the Browser API.
 
 ### [v0.31](https://github.com/realityforge/router-fu/tree/v0.31) (2020-06-23) · [Full Changelog](https://github.com/realityforge/router-fu/compare/v0.30...v0.31)